requirements.txt
telethon>=1.31.1
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
